+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add `INSPECT_DISABLE_MODEL_API` environment variable for disabling all Model APIs save for mockllm.
 - Add optional `tool_call_id` param to `ModelOutput.for_tool_call()`.
 - Don't apply a reducer if there is only a single epoch (previously the reducer would be run against the single epoch value, which had side effects on score values)
+- Support all JSON and CSV dataset arguments in `file_dataset()` function.
 
 ## v0.3.42 (23 October 2024)